+++ src/main/java/roomescape/common/PageController.java
@@ -1,4 +1,4 @@
-package roomescape.global.common;
+package roomescape.common;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
