+++ src/main/java/roomescape/config/WebConfig.java
@@ -1,14 +1,13 @@
-package roomescape.global.config;
+package roomescape.config;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.ssl.DefaultSslBundleRegistry;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.global.auth.AdminAuthInterceptor;
-import roomescape.global.auth.LoginMemberArgumentResolver;
+import roomescape.auth.AdminAuthInterceptor;
+import roomescape.auth.LoginMemberArgumentResolver;
 
 @Configuration
 @RequiredArgsConstructor
@@ -16,14 +15,12 @@ public class WebConfig implements WebMvcConfigurer {
 
     private final AdminAuthInterceptor adminAuthInterceptor;
     private final LoginMemberArgumentResolver loginMemberArgumentResolver;
-    private final DefaultSslBundleRegistry sslBundleRegistry;
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(adminAuthInterceptor)
                 .addPathPatterns("/**")
                 .excludePathPatterns("/login", "/signup");
-
     }
 
     @Override
