+++ src/main/java/roomescape/exception/CustomException.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.exception;
 
 import lombok.Getter;
