+++ src/main/resources/schema.sql
@@ -1,42 +1,3 @@
-CREATE TABLE IF NOT EXISTS time
-(
-    id         BIGINT      NOT NULL AUTO_INCREMENT,
-    time_value VARCHAR(20) NOT NULL,
-    deleted    BOOLEAN     NOT NULL DEFAULT FALSE,
-    PRIMARY KEY (id)
-);
-
-CREATE TABLE IF NOT EXISTS theme
-(
-    id          BIGINT       NOT NULL AUTO_INCREMENT,
-    name        VARCHAR(255) NOT NULL,
-    description VARCHAR(255) NOT NULL,
-    deleted     BOOLEAN      NOT NULL DEFAULT FALSE,
-    PRIMARY KEY (id)
-);
-
-CREATE TABLE IF NOT EXISTS member
-(
-    id       BIGINT              NOT NULL AUTO_INCREMENT,
-    name     VARCHAR(255)        NOT NULL,
-    email    VARCHAR(255) UNIQUE NOT NULL,
-    password VARCHAR(255)        NOT NULL,
-    role     VARCHAR(255)        NOT NULL,
-    PRIMARY KEY (id)
-);
-
-CREATE TABLE IF NOT EXISTS reservation
-(
-    id       BIGINT       NOT NULL AUTO_INCREMENT,
-    date     VARCHAR(255) NOT NULL,
-    name     VARCHAR(255) NOT NULL,
-    time_id  BIGINT,
-    theme_id BIGINT,
-    PRIMARY KEY (id),
-    FOREIGN KEY (time_id) REFERENCES time (id),
-    FOREIGN KEY (theme_id) REFERENCES theme (id)
-);
-
 INSERT INTO member (name, email, password, role)
 VALUES ('어드민', 'admin@email.com', 'password', 'ADMIN'),
        ('브라운', 'brown@email.com', 'password', 'USER');
